Move things around to make the hierarchy more revealing
