+++ src/backends/penta/PentabarfParser.ts
@@ -15,10 +15,10 @@ limitations under the License.
 */
 
 import * as parser from 'fast-xml-parser';
-import { IAuditorium, IConference, IInterestRoom, IPerson, ITalk, Role } from "../models/schedule";
+import { IAuditorium, IConference, IInterestRoom, IPerson, ITalk, Role } from "../../models/schedule";
 import * as moment from "moment";
-import { RoomKind } from "../models/room_kinds";
-import config from "../config";
+import { RoomKind } from "../../models/room_kinds";
+import config from "../../config";
 
 function arrayLike<T>(val: T | T[]): T[] {
     if (Array.isArray(val)) return val;
